🔧 fix: accessToken 발급 로직 통합
