+++ BE/src/koreaInvestment/korea.investment.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { KoreaInvestmentService } from './korea.investment.service';
+
+@Module({
+  imports: [],
+  controllers: [],
+  providers: [KoreaInvestmentService],
+  exports: [KoreaInvestmentService],
+})
+export class KoreaInvestmentModule {}
