+++ BE/src/stock/index/stock.index.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { StockIndexController } from './stock.index.controller';
 import { StockIndexService } from './stock.index.service';
+import { KoreaInvestmentModule } from '../../koreaInvestment/korea.investment.module';
 
 @Module({
-  imports: [],
+  imports: [KoreaInvestmentModule],
   controllers: [StockIndexController],
   providers: [StockIndexService],
   exports: [StockIndexService],
