+++ BE/src/stock/topfive/stock.topfive.module.ts
@@ -2,9 +2,10 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { StockTopfiveController } from './stock.topfive.controller';
 import { StockTopfiveService } from './stock.topfive.service';
+import { KoreaInvestmentModule } from '../../koreaInvestment/korea.investment.module';
 
 @Module({
-  imports: [ConfigModule],
+  imports: [ConfigModule, KoreaInvestmentModule],
   controllers: [StockTopfiveController],
   providers: [StockTopfiveService],
 })
