+++ BE/src/stock/topfive/stock.topfive.service.ts
@@ -9,11 +9,10 @@ import {
   StockApiResponse,
 } from './interface/stock.topfive.interface';
 import { MarketType } from '../enum/MarketType';
+import { KoreaInvestmentService } from '../../koreaInvestment/korea.investment.service';
 
 @Injectable()
 export class StockTopfiveService {
-  private accessToken: string;
-  private tokenExpireTime: Date;
   private readonly koreaInvestmentConfig: {
     appKey: string;
     appSecret: string;
@@ -22,7 +21,10 @@ export class StockTopfiveService {
 
   private readonly logger = new Logger();
 
-  constructor(private readonly config: ConfigService) {
+  constructor(
+    private readonly config: ConfigService,
+    private readonly koreaInvestmentService: KoreaInvestmentService,
+  ) {
     this.koreaInvestmentConfig = {
       appKey: this.config.get<string>('KOREA_INVESTMENT_APP_KEY'),
       appSecret: this.config.get<string>('KOREA_INVESTMENT_APP_SECRET'),
@@ -30,30 +32,9 @@ export class StockTopfiveService {
     };
   }
 
-  private async getAccessToken() {
-    // accessToken이 유효한 경우
-    if (this.accessToken && this.tokenExpireTime > new Date()) {
-      return this.accessToken;
-    }
-
-    const response = await axios.post(
-      `${this.koreaInvestmentConfig.baseUrl}/oauth2/tokenP`,
-      {
-        grant_type: 'client_credentials',
-        appkey: this.koreaInvestmentConfig.appKey,
-        appsecret: this.koreaInvestmentConfig.appSecret,
-      },
-    );
-
-    this.accessToken = response.data.access_token;
-    this.tokenExpireTime = new Date(Date.now() + +response.data.expires_in);
-
-    return this.accessToken;
-  }
-
   private async requestApi(params: StockRankigRequestDto) {
     try {
-      const token = await this.getAccessToken();
+      const token = await this.koreaInvestmentService.getAccessToken();
 
       const response = await axios.get<StockApiResponse>(
         `${this.koreaInvestmentConfig.baseUrl}/uapi/domestic-stock/v1/ranking/fluctuation`,
